extracted logic of blake into blake_core
