+++ arithmetization/src/test/examples/rust/src/blake_core.rs
@@ -1,15 +1,4 @@
-#![no_std]
-#![no_main]
-
-/// Blake2b‑F compression function (used in EIP‑152 precompile, zkVMs, etc.)
-///
-/// Reference: RFC 7693, plus the Java gist:
-/// https://gist.github.com/DavePearce/fca4c7fcfac840dc362b1c907d672093
-
-use core::convert::TryInto;
-use core::result::Result;
-use core::result::Result::Err;
-use core::result::Result::Ok;
+// This logic is used in both the `blake.rs` (freestanding) and `blake_from_main.rs` (assuming underlying OS)
 
 /// Initial state vector (IV) for Blake2b
 const IV: [u64; 8] = [
@@ -201,72 +190,4 @@ pub fn blake2b_f_eip152(input: &[u8]) -> Result<[u8; 64], &'static str> {
     }
 
     Ok(out)
-}
-
-core::arch::global_asm!(
-    ".global _start",
-    "_start:",
-    "li sp, 0x087fffff", // set stack pointer to a known memory region
-    "call main",
-);
-
-#[no_mangle]
-fn main() -> ! {
-    let test_vector_4 = [
-        "0000000048c9bdf267e6096a3ba7ca8485ae67bb2bf894fe72f36e3cf1361d5f3af54fa5d182e6ad7f520e511f6c3e2b8c68059b6bbd41fbabd9831f79217e1319cde05b61626300000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000300000000000000000000000000000001",
-        "08c9bcf367e6096a3ba7ca8485ae67bb2bf894fe72f36e3cf1361d5f3af54fa5d282e6ad7f520e511f6c3e2b8c68059b9442be0454267ce079217e1319cde05b" ];
-
-    let test_vector_5 =["0000000c48c9bdf267e6096a3ba7ca8485ae67bb2bf894fe72f36e3cf1361d5f3af54fa5d182e6ad7f520e511f6c3e2b8c68059b6bbd41fbabd9831f79217e1319cde05b61626300000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000300000000000000000000000000000001","ba80a53f981c4d0d6a2797b69f12f6e94c212f14685ac4b74b12bb6fdbffa2d17d87c5392aab792dc252d5de4533cc9518d38aa8dbf1925ab92386edd4009923"];
-
-    let test_vector_6 = [
-        "0000000c48c9bdf267e6096a3ba7ca8485ae67bb2bf894fe72f36e3cf1361d5f3af54fa5d182e6ad7f520e511f6c3e2b8c68059b6bbd41fbabd9831f79217e1319cde05b61626300000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000300000000000000000000000000000000",
-        "75ab69d3190a562c51aef8d88f1c2775876944407270c42c9844252c26d2875298743e7f6d5ea2f2d3e8d226039cd31b4e426ac4f2d3d666a610c2116fde4735" ];
-
-    let test_vector_7 = [
-        "0000000148c9bdf267e6096a3ba7ca8485ae67bb2bf894fe72f36e3cf1361d5f3af54fa5d182e6ad7f520e511f6c3e2b8c68059b6bbd41fbabd9831f79217e1319cde05b61626300000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000300000000000000000000000000000001",
-        "b63a380cb2897d521994a85234ee2c181b5f844d2c624c002677e9703449d2fba551b3a8333bcdf5f2f7e08993d53923de3d64fcc68c034e717b9293fed7a421" ];
-
-    let _test_vector_8 = [
-        "ffffffff48c9bdf267e6096a3ba7ca8485ae67bb2bf894fe72f36e3cf1361d5f3af54fa5d182e6ad7f520e511f6c3e2b8c68059b6bbd41fbabd9831f79217e1319cde05b61626300000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000300000000000000000000000000000001",
-        "fc59093aafa9ab43daae0e914c57635c5402d8e3d2130eb9b3cc181de7f0ecf9b22bf99a7815ce16419e200e01846e6b5df8cc7703041bbceb571de6631d2615" ];
-
-    // Note: test vector 8 is not included for now as number of rounds is 0xffffffff
-    let test_vectors = [test_vector_4, test_vector_5, test_vector_6, test_vector_7];
-
-    for i in 0..test_vectors.len() {
-        let input = hex_to_input(test_vectors[i][0]);
-        let expected = hex_to_expected(test_vectors[i][1]);
-
-        let _code = match blake2b_f_eip152(&input) {
-            Ok(result) if result == expected.as_slice() => 0, // success
-            Ok(_) => 1,                                       // wrong result
-            Err(_) => 2,                                      // compression failed
-        };
-
-        // println!(
-        //     "Test vector {}: {}",
-        //     i + 4,
-        //     if code == 0 { "PASS" } else { "FAIL" }
-        // );
-    }
-
-    exit(0)
-}
-
-fn exit(code: i32) -> ! {
-    unsafe {
-        core::arch::asm!(
-            "mv a0, {0}",  // exit code
-            "li a7, 93",   // syscall number for exit
-            "ecall",
-            in(reg) code,
-            options(noreturn)
-        );
-    }
-}
-
-// required by the compiler
-#[panic_handler]
-fn panic(_: &core::panic::PanicInfo) -> ! {
-    exit(3);
 }
